Preserve file data when power is unexpectedly lost
This is achieved by increasing the filesize in the directory each time a
cluster is appended, then trimming the filesize to the logical length of
the file when it is finally closed. 

This allows nearly all data written to the file to be recovered on
powerloss (only the last 512 byte sector should be lost) at the expense
of including garbage at the end of the file that was not part of the
logical file.